counterfeit_detection/requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3

# Database (compatible Python 3.8)
sqlalchemy==2.0.23

# Data Processing (compatible Python 3.8)
pandas==2.0.3
numpy==1.24.4

# Visualization & Dashboard
streamlit==1.28.0
plotly==5.17.0

# AI/ML Models (for image similarity - compatible Python 3.8)
torch==1.13.1
torchvision==0.14.1

# Image Processing
pillow==10.0.1
opencv-python==4.8.1.78

# Utilities
python-dotenv==1.0.0
schedule==1.2.0

# Logging
loguru==0.7.2
